// src/code/model/operation.d.ts
export declare type OpPermItem = {
    role: "read" | "write" | "assignonly";
    teamid: string;
    zone: number;
};